requirements.txt
streamlit
openai
PyPDF2
python-docx
scikit-learn
spacy
pandas
python-multipart
gunicorn